AOJ ITP1_10_A to_iをto_fに修正(AC) (#50)

+++ AOJ/ITP1/10_A.rb
@@ -1,3 +1,3 @@
-x1,y1,x2,y2 = gets.split.map(&:to_i)
+x1,y1,x2,y2 = gets.split.map(&:to_f)
 
 p Math.sqrt(((x2 - x1) ** 2) + ((y2 - y1) ** 2))
